# Merge conflicts resolved.

+++ memi/memi-web/src/main/java/uk/ac/ebi/interpro/metagenomics/memi/springmvc/modelbuilder/HomePageViewModelBuilder.java
@@ -79,9 +79,9 @@ public class HomePageViewModelBuilder extends AbstractBiomeViewModelBuilder<Home
         final int publicRunCount = runDAO.countAllPublic();
         final int privateRunCount = runDAO.countAllPrivate();
 
-        final Map<String, Integer> experimentCountMap = runDAO.retrieveRunCountsGroupedByExperimentType(3);
-        final Map<String, Integer> transformedExperimentCountMap = transformMap(experimentCountMap);
-        final Integer numOfDataSets = getNumOfDataSets(experimentCountMap);
+        final Map<String, Long> experimentCountMap = runDAO.retrieveRunCountsGroupedByExperimentType(3);
+        final Map<String, Long> transformedExperimentCountMap = transformMap(experimentCountMap);
+        final Long numOfDataSets = getNumOfDataSets(experimentCountMap);
 
         List<Study> studies = null;
         // If case: if nobody is logged in
@@ -162,8 +162,8 @@ public class HomePageViewModelBuilder extends AbstractBiomeViewModelBuilder<Home
         return result;
     }
 
-    private Integer getNumOfDataSets(Map<String, Integer> experimentCountMap) {
-        Integer result = 0;
+    private Long getNumOfDataSets(Map<String, Long> experimentCountMap) {
+        Long result = 0L;
         for (String key : experimentCountMap.keySet()) {
             result += experimentCountMap.get(key);
         }
